Finished code splitting
