+++ src/main/resources/assets/src/detailsPage.js
@@ -1,0 +1,15 @@
+import React from 'react';
+
+export default class Details extends React.Component {
+    constructor(props) {
+        super(props);
+    }
+
+    render() {
+        return (
+            <div>
+                <p>{"Here's your appointment"}</p>
+            </div>
+        );
+    }
+}
